Add files via upload
These two files control everything right now. The database "DegreeViz.db" has all of the tables for Courses, Grades and Users. When you run "Authenticate.py" make sure the DegreeViz.db is saved in the same directory (best bet is to just pull from this repository and run in VS).
If this becomes a problem we will have to store a .db locally.